requirements.txt
streamlit>=1.42.0
python-dotenv>=1.0.1
azure-ai-inference==1.0.0b8
azure-core>=1.29.5